feat: pass LLM proxy URL and token to MCP servers
This change will give MCP servers access to the LLM proxy in Obot. Since
MCP servers are scoped to the project, then so are their LLM requests.
Meaning that, if the user has set a model at the thread level, LLM
requests from MCP servers will not see this and will use the
project-level configuration.

Signed-off-by: Donnie Adams <donnie@acorn.io>