Chore: Allow UI to hit remote API for dev (#5380)
In local development it is useful to be able to hit a remote instance of
the Obot API, like hitting our "main" environment so that you can
development the frontend against a backend with a lot more data.

This change makes that possible. You need to set two environment
variables when running npm run dev:
- VITE_API_TARGET=https://<hostname>
- VITE_API_TOKEN=<token>

A few notes on those:
- Note that the API target does NOT include /api at the end, just the
  hostname
- The token will be set as the Authorization Bearer token and is used to
  bypass OAuth. You can obtain a token through the Obot CLI with:
`OBOT_BASE_URL=https://<hostname>/api obot token`
Note that you do include the /api when getting a token this way.

Signed-off-by: Craig Jellick <craig@obot.ai>